Add ext-mongodb to composer.json, event store now require collection, some cleanup
